refactor: Rental Repository 를 port/adaptor 구조로 분리
