--- admin/src/main/java/kernel360/ckt/admin/infra/RentalRepository.java
@@ -1,36 +1,0 @@
-package kernel360.ckt.admin.infra;
-
-import java.time.LocalDateTime;
-import java.util.List;
-import java.util.Optional;
-import kernel360.ckt.core.domain.entity.RentalEntity;
-import kernel360.ckt.core.domain.entity.VehicleEntity;
-import kernel360.ckt.core.domain.enums.RentalStatus;
-import org.springframework.data.domain.Page;
-import org.springframework.data.domain.Pageable;
-
-public interface RentalRepository {
-    RentalEntity save(RentalEntity rental);
-
-    List<RentalEntity> findOverlappingRentalsByVehicleAndStatuses(
-        VehicleEntity vehicle,
-        List<RentalStatus> statuses,
-        LocalDateTime pickupAt,
-        LocalDateTime returnAt
-    );
-
-    Page<RentalEntity> search(
-        Long companyId,
-        RentalStatus status,
-        String keyword,
-        LocalDateTime startAt,
-        LocalDateTime endAt,
-        Pageable pageable
-    );
-
-    Optional<RentalEntity> findById(Long id);
-
-    List<RentalEntity> findRentalsByStatus(RentalStatus status);
-
-    long countVehiclesByStatus(RentalStatus status);
-}
