+++ admin/src/main/java/kernel360/ckt/admin/infra/jpa/RentalJpaRepository.java
@@ -6,14 +6,16 @@ import java.util.Optional;
 import kernel360.ckt.core.domain.entity.RentalEntity;
 import kernel360.ckt.core.domain.entity.VehicleEntity;
 import kernel360.ckt.core.domain.enums.RentalStatus;
-import kernel360.ckt.admin.infra.RentalRepository;
+import kernel360.ckt.admin.application.port.RentalRepository;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.jpa.repository.EntityGraph;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.query.Param;
+import org.springframework.stereotype.Repository;
 
+@Repository
 public interface RentalJpaRepository extends JpaRepository<RentalEntity, Long>, RentalRepository {
     @Query("""
         SELECT r FROM RentalEntity r
@@ -44,7 +46,7 @@ public interface RentalJpaRepository extends JpaRepository<RentalEntity, Long>, 
         AND (:endAt IS NULL OR r.returnAt < :endAt)
         ORDER BY r.pickupAt DESC
     """)
-    Page<RentalEntity> search(
+    Page<RentalEntity> findAll(
         @Param("companyId") Long companyId,
         @Param("status") RentalStatus status,
         @Param("keyword") String keyword,
